requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
asgiref==3.7.2
attrs==24.2.0
autopep8==2.3.1
discord==2.3.2
discord.py==2.4.0
Django==5.0
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
pycodestyle==2.12.1
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
yarl==1.17.2
